backend/requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database ORM
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.12.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# OpenAI Agents SDK
openai>=1.0.0

# MCP (Model Context Protocol) SDK
# Note: Install official MCP SDK when available
# For now, we'll implement MCP tools as Python functions
# mcp-server-python>=0.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.1
Jinja2>=3.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
ruff>=0.0.285
mypy>=1.5.0
